chore: Move kafka example to own package
