+++ springwolf-examples/springwolf-amqp-example/src/test/java/io/github/stavshamir/springwolf/example/amqp/SpringwolfAmqpExampleApplicationTest.java
@@ -8,7 +8,7 @@ import org.springframework.context.ApplicationContext;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 @SpringBootTest
-class SpringwolfExampleApplicationTest {
+class SpringwolfAmqpExampleApplicationTest {
 
     @Autowired
     private ApplicationContext context;
