+++ springwolf-examples/springwolf-cloud-stream-example/src/test/java/io/github/stavshamir/springwolf/example/cloudstream/SpringContextTest.java
@@ -16,7 +16,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 public class SpringContextTest {
 
-    @SpringBootTest(classes = SpringwolfExampleApplication.class)
+    @SpringBootTest(classes = SpringwolfCloudstreamExampleApplication.class)
     @EmbeddedKafka(partitions = 1, brokerProperties = {"listeners=PLAINTEXT://localhost:29092", "port=29092"})
     @Nested
     @DirtiesContext
@@ -35,7 +35,7 @@ public class SpringContextTest {
         }
     }
 
-    @SpringBootTest(classes = SpringwolfExampleApplication.class)
+    @SpringBootTest(classes = SpringwolfCloudstreamExampleApplication.class)
     @EmbeddedKafka(partitions = 1, brokerProperties = {"listeners=PLAINTEXT://localhost:29092", "port=29092"})
     @Nested
     @DirtiesContext
