+++ springwolf-examples/springwolf-kafka-example/src/main/java/io/github/stavshamir/springwolf/example/kafka/dtos/ExamplePayloadDto.java
@@ -1,4 +1,4 @@
-package io.github.stavshamir.springwolf.example.amqp.dtos;
+package io.github.stavshamir.springwolf.example.kafka.dtos;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.EqualsAndHashCode;
