+++ springwolf-examples/springwolf-kafka-example/src/main/resources/application.properties
@@ -27,7 +27,7 @@ spring.kafka.consumer.properties.spring.json.trusted.packages=io.github.stavsham
 #########
 # Springwolf configuration
 springwolf.enabled=true
-springwolf.docket.base-package=io.github.stavshamir.springwolf.example
+springwolf.docket.base-package=io.github.stavshamir.springwolf.example.kafka
 springwolf.docket.info.title=${spring.application.name}
 springwolf.docket.info.version=1.0.0
 springwolf.docket.info.description=Springwolf example project to demonstrate springwolfs abilities
